% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{as.matrix.mrts}
\alias{as.matrix.mrts}
\title{Internal function: Remove attributes of mrts}
\usage{
as.matrix.mrts(x, ...)
}
\arguments{
\item{x}{A mrts object}

\item{...}{Not used directly}
}
\value{
A matrix object
}
\description{
Internal function: Remove attributes of mrts
}
\keyword{internal}
