% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calculateLatticeKrigBasis}
\alias{calculateLatticeKrigBasis}
\title{Internal function: A modifier of LatticeKrig::LKrig.basis}
\usage{
calculateLatticeKrigBasis(x1, LKinfo)
}
\arguments{
\item{x1}{A matrix}

\item{LKinfo}{LKinfo object}
}
\value{
A matrix
}
\description{
Internal function: A modifier of LatticeKrig::LKrig.basis
}
\keyword{internal}
