% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calculateLogDeterminant}
\alias{calculateLogDeterminant}
\title{Internal function: calculate the log determinant for the likelihood use.}
\usage{
calculateLogDeterminant(R, L, K)
}
\arguments{
\item{R}{A p x p matrix}

\item{L}{A p x K matrix}

\item{K}{A numeric}
}
\value{
A numeric
}
\description{
Internal function: calculate the log determinant for the likelihood use.
}
\keyword{internal}
