% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkObject.R
\name{calculateSARForOneDimLocation}
\alias{calculateSARForOneDimLocation}
\title{Internal function: A modifier of LatticeKrig::LKrigSAR.LKInterval}
\usage{
calculateSARForOneDimLocation(LKinfo, level)
}
\arguments{
\item{LKinfo}{LKinfo object}

\item{level}{an integer}
}
\value{
A list
}
\description{
Internal function: A modifier of LatticeKrig::LKrigSAR.LKInterval
}
\keyword{internal}
