% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calculateSARForTowDimLocation}
\alias{calculateSARForTowDimLocation}
\title{Internal function: A modifier of LatticeKrig::LKrigSAR.LKBox}
\usage{
calculateSARForTowDimLocation(LKinfo, level)
}
\arguments{
\item{LKinfo}{LKinfo object}

\item{level}{an integer}
}
\value{
A list
}
\description{
Internal function: A modifier of LatticeKrig::LKrigSAR.LKBox
}
\keyword{internal}
