% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{cMLElk}
\alias{cMLElk}
\title{Internal function: cMLElk}
\usage{
cMLElk(Fk, data, Depsilon, wSave = FALSE, DfromLK, vfixed = NULL)
}
\arguments{
\item{Fk}{An \emph{n} by \emph{K} matrix of basis function values with
each column being a basis function taken values at \code{loc}.}

\item{data}{An \emph{n} by \emph{T} data matrix (NA allowed) with}

\item{Depsilon}{An \emph{n} by \emph{n} diagonal matrix.}

\item{wSave}{A logic.}

\item{DfromLK}{An \emph{n} by \emph{n} matrix}

\item{vfixed}{A numeric.}
}
\value{
A list.
}
\description{
Internal function: cMLElk
}
\keyword{internal.}
