% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{cMLEsp}
\alias{cMLEsp}
\title{Internal function: cMLEsp}
\usage{
cMLEsp(Fk, data, Depsilon, wSave = FALSE)
}
\arguments{
\item{Fk}{An \emph{n} by \emph{K} matrix of basis function values with
each column being a basis function taken values at \code{loc}.}

\item{data}{An \emph{n} by \emph{T} data matrix (NA allowed) with}

\item{Depsilon}{An \emph{n} by \emph{n} diagonal matrix.}

\item{wSave}{A logic.}
}
\value{
A list.
}
\description{
Internal function: cMLEsp
}
\keyword{internal.}
