% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeMrtsRcpp}
\alias{computeMrtsRcpp}
\title{Internal function: Compute MRTS method}
\usage{
computeMrtsRcpp(s, xobs_diag, k)
}
\arguments{
\item{s}{A location matrix}

\item{xobs_diag}{A matrix of observations}

\item{k}{A rank}
}
\value{
A list of objects
\item{X}{A matrix}
\item{UZ}{A matrix}
\item{BBBH}{A matrix}
\item{nconst}{A vector of column means}
}
\description{
Internal function: Compute MRTS method
}
\author{
Wen-Ting Wang
}
\keyword{internal}
