% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{computeNegativeLikelihood}
\alias{computeNegativeLikelihood}
\title{Internal function: compute a negative likelihood}
\usage{
computeNegativeLikelihood(
  nrow_Fk,
  ncol_Fk,
  n,
  p,
  matrix_JSJ,
  sample_covariance_trace,
  vfixed = NULL,
  ldet = 0
)
}
\arguments{
\item{nrow_Fk}{An integer. The number of rows of Fk.}

\item{ncol_Fk}{An integer. The number of columns of Fk.}

\item{n}{An integer. Sample size.}

\item{p}{A positive integers. The number of columns of data.}

\item{matrix_JSJ}{A multiplication matrix}

\item{sample_covariance_trace}{A positive numeric.}

\item{vfixed}{A numeric}

\item{ldet}{A numeric. A log determinant.}
}
\value{
A list.
}
\description{
Internal function: compute a negative likelihood
}
\keyword{internal.}
