% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{computeProjectionMatrix}
\alias{computeProjectionMatrix}
\title{Internal function: maximum likelihood estimate with the likelihood}
\usage{
computeProjectionMatrix(Fk1, Fk2, data, S = NULL)
}
\arguments{
\item{Fk1}{An \emph{n} by \emph{K} matrix}

\item{Fk2}{An \emph{n} by \emph{K} matrix}

\item{data}{An \emph{n} by \emph{T} data matrix}

\item{S}{An \emph{n} by \emph{n} matrix}
}
\value{
A list.
}
\description{
Internal function: maximum likelihood estimate with the likelihood
}
\keyword{internal.}
