% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convertToPositiveDefinite}
\alias{convertToPositiveDefinite}
\title{Internal function: convert a matrix to positive definite}
\usage{
convertToPositiveDefinite(mat)
}
\arguments{
\item{mat}{A matrix or a dataframe}
}
\value{
A positive-definite matrix
}
\description{
Internal function: convert a matrix to positive definite
}
\keyword{internal}
