% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenDecompose}
\alias{eigenDecompose}
\title{Internal function: Eigen-decompose a matrix}
\usage{
eigenDecompose(matrix)
}
\arguments{
\item{matrix}{A matrix}
}
\value{
A list of objects
\item{value}{A vector of eigenvalues}
\item{vector}{A matrix of eigenvectors}
}
\description{
Internal function: Eigen-decompose a matrix
}
\keyword{internal}
