% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{eigenDecomposeInDecreasingOrder}
\alias{eigenDecomposeInDecreasingOrder}
\title{Internal function: eigen-decomposition in decreasing order}
\usage{
eigenDecomposeInDecreasingOrder(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A list.
}
\description{
Internal function: eigen-decomposition in decreasing order
}
\keyword{internal.}
