% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{EM0miss}
\alias{EM0miss}
\title{Internal function: EM0miss}
\usage{
EM0miss(
  Fk,
  data,
  Depsilon,
  maxit,
  avgtol,
  wSave = FALSE,
  external = FALSE,
  DfromLK = NULL,
  vfixed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{Fk}{An \emph{n} by \emph{K} matrix of basis function values with
each column being a basis function taken values at \code{loc}.}

\item{data}{An \emph{n} by \emph{T} data matrix (NA allowed) with}

\item{Depsilon}{An \emph{n} by \emph{n} diagonal matrix.}

\item{maxit}{An integer for the maximum number of iterations.}

\item{avgtol}{A numeric for average tolerance.}

\item{wSave}{A logic.}

\item{external}{A logic.}

\item{DfromLK}{An \emph{n} by \emph{n} matrix}

\item{vfixed}{A numeric.}

\item{verbose}{A logic. Print a useful information.}
}
\value{
A list.
}
\description{
Internal function: EM0miss
}
\keyword{internal.}
