% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{estimateEta}
\alias{estimateEta}
\title{Internal function: estimate eta parameter}
\usage{
estimateEta(d, s, v)
}
\arguments{
\item{d}{An array of nonnegative values.}

\item{s}{A positive numeric.}

\item{v}{A positive numeric.}
}
\value{
A numeric.
}
\description{
Internal function: estimate eta parameter
}
\keyword{internal.}
