% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{estimateV}
\alias{estimateV}
\title{Internal function: solve v parameter}
\usage{
estimateV(d, s, sample_covariance_trace, n)
}
\arguments{
\item{d}{An array of nonnegative values.}

\item{s}{A positive numeric.}

\item{sample_covariance_trace}{A positive numeric.}

\item{n}{An integer. Sample size.}
}
\value{
A numeric.
}
\description{
Internal function: solve v parameter
}
\keyword{internal.}
