% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fetchNonZeroIndexs}
\alias{fetchNonZeroIndexs}
\title{Internal function: convert to a sparse matrix}
\usage{
fetchNonZeroIndexs(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
An array of indeces
}
\description{
Internal function: convert to a sparse matrix
}
\keyword{internal}
