% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fetchSystemRam}
\alias{fetchSystemRam}
\title{Internal function: fetch the system ram}
\usage{
fetchSystemRam(os)
}
\arguments{
\item{os}{An OS name with type}
}
\value{
An integer
}
\description{
Internal function: fetch the system ram
}
\keyword{internal}
