% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getInverseSquareRootMatrix}
\alias{getInverseSquareRootMatrix}
\title{Internal function: Inverse square root matrix of  A^t * B}
\usage{
getInverseSquareRootMatrix(left_matrix, right_matrix)
}
\arguments{
\item{left_matrix}{A matrix}

\item{right_matrix}{A matrix}
}
\value{
A matrix
}
\description{
Internal function: Inverse square root matrix of  A^t * B
}
\author{
Wen-Ting Wang
}
\keyword{internal}
