% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSquareRootMatrix}
\alias{getSquareRootMatrix}
\title{Internal function: A wrapper function of 'MatrixBase::sqrt()'}
\usage{
getSquareRootMatrix(matrix)
}
\arguments{
\item{matrix}{A matrix}
}
\value{
A matrix
}
\description{
Internal function: A wrapper function of 'MatrixBase::sqrt()'
}
\keyword{internal}
