% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkObject.R
\name{initializeLKnFRK}
\alias{initializeLKnFRK}
\title{Internal function: initializeLKnFRK}
\usage{
initializeLKnFRK(
  data,
  location,
  nlevel = 3,
  weights = NULL,
  n.neighbor = 3,
  nu = 1
)
}
\arguments{
\item{data}{\emph{n} by \emph{T} data matrix (NA allowed) with
\eqn{z[t]} as the \emph{t}-th column.}

\item{location}{\emph{n} by \emph{d} matrix of coordinates corresponding to \emph{n} locations.}

\item{weights}{An \emph{n} by \emph{n} diagonal matrix.}

\item{n.neighbor}{number of neighbors to be used in the "fast" imputation method. Default is 3.}

\item{nu}{An integer.}

\item{nlelve}{An integer.}
}
\value{
list
}
\description{
Internal function: initializeLKnFRK
}
\keyword{internal}
