% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{invCz}
\alias{invCz}
\title{Internal function: internal matrix calcuation}
\usage{
invCz(R, L, z)
}
\arguments{
\item{R}{A p x p matrix}

\item{L}{A p x K matrix}

\item{z}{An array with length p or 1 x p matrix}
}
\value{
A 1 x p matrix
}
\description{
Internal function: internal matrix calcuation
}
\keyword{internal}
