% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{isDiagonal}
\alias{isDiagonal}
\title{Internal function: check if a numeric-like object is diagonal}
\usage{
isDiagonal(object)
}
\arguments{
\item{object}{An R object}
}
\value{
logical
}
\description{
Internal function: check if a numeric-like object is diagonal
}
\keyword{internal}
