% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{LKpeon}
\alias{LKpeon}
\title{Internal function: LKpeon}
\usage{
LKpeon(
  M,
  s,
  Fk,
  basis,
  weight,
  phi1,
  phi0,
  Q,
  lambda,
  phi0P,
  L = NULL,
  data = NULL,
  only.wlk = FALSE,
  only.se = FALSE
)
}
\arguments{
\item{M}{A \emph{K} by \emph{K} symmetric matrix.}

\item{s}{A scalar.}

\item{Fk}{An \emph{n} by \emph{K} matrix of basis function values with
each column being a basis function taken values at \code{loc}.}

\item{basis}{An \emph{n} by \emph{K} matrix}

\item{weight}{A vector}

\item{phi1}{A matrix produced by LatticeKrig basis fucntions}

\item{phi0}{A matrix produced by LatticeKrig basis fucntions with new locations}

\item{Q}{A precision matrix}

\item{lambda}{A scalar.}

\item{phi0P}{A projection matrix of phi1}

\item{L}{An \emph{n} by \emph{K} matrix}

\item{data}{An \emph{n} by \emph{T} data matrix}

\item{only.wlk}{A logic.}

\item{only.se}{A logic.}
}
\value{
A list.
}
\description{
Internal function: LKpeon
}
\keyword{internal}
