% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{LKrigSetupWrapper}
\alias{LKrigSetupWrapper}
\title{Internal function: A wrapper for LatticeKrig::LKrigSetup}
\usage{
LKrigSetupWrapper(
  x = NULL,
  nlevel = NULL,
  alpha = NA,
  a.wght = NA,
  NC = NULL,
  lambda = NA,
  LKGeometry = "LKRectangle",
  ...
)
}
\arguments{
\item{x}{Spatial locations that define the spatial domain for prediction.}

\item{nlevel}{Number of levels in multi-resolution.}

\item{alpha}{A vector of length nlevel with the relative variances for the different multi- resolution levels.}

\item{a.wght}{The correlation range in the SAR model.}

\item{NC}{The maximum number of lattice grid points for a spatial coordinate and at the coarsest level of resolution.}

\item{lambda}{A smoothing parameter.}

\item{LKGeometry}{A text string that gives the names of the model geometry.}
}
\value{
list
}
\description{
Internal function: A wrapper for LatticeKrig::LKrigSetup
}
\keyword{internal}
