% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{logDeterminant}
\alias{logDeterminant}
\title{Internal function: log-determinant of a sqaure matrix}
\usage{
logDeterminant(mat)
}
\arguments{
\item{mat}{A sqaure matrix.}
}
\value{
A numeric.
}
\description{
Internal function: log-determinant of a sqaure matrix
}
\keyword{internal}
