% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkObject.R
\name{normalizeBasis}
\alias{normalizeBasis}
\title{Internal function: A modifier of LatticeKrig::LKrigNormalizeBasis}
\usage{
normalizeBasis(LKinfo, level, PHI)
}
\arguments{
\item{LKinfo}{LKinfo object}

\item{x1}{A matrix}
}
\value{
A matrix
}
\description{
Internal function: A modifier of LatticeKrig::LKrigNormalizeBasis
}
\keyword{internal}
