% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFRK.R
\name{predict.mrts}
\alias{predict.mrts}
\title{Multi-Resolution Thin-plate Spline Basis Functions}
\usage{
\method{predict}{mrts}(object, newx, ...)
}
\arguments{
\item{object}{object produced from calling mrts.}

\item{newx}{an \emph{n} by \emph{d} matrix of coordinates corresponding to \emph{n} locations.}

\item{...}{not used but needed for the S3 generic/method compatibility.}
}
\value{
an \emph{n} by \emph{k} matrix of the \emph{k} basis function in \code{object} taken values at \code{newx}.
}
\description{
Evaluate multi-resolution thin-plate spline basis  functions at given locations.
This function provides a generic prediction method for \code{mrts} objects, 
in a similar way as \code{predict.ns} and \code{predict.bs} in \code{splines} package.
}
\seealso{
\code{\link{mrts}}
}
\author{
ShengLi Tzeng and Hsin-Cheng Huang.
}
