% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictMrtsRcpp}
\alias{predictMrtsRcpp}
\title{Internal function: Predict on new locatoins by MRTS method}
\usage{
predictMrtsRcpp(s, xobs_diag, s_new, k)
}
\arguments{
\item{s}{A location matrix}

\item{xobs_diag}{A matrix of observations}

\item{s_new}{A new location matrix}

\item{k}{A rank}
}
\value{
A list of objects
\item{X}{A matrix}
\item{UZ}{A matrix}
\item{BBBH}{A matrix}
\item{nconst}{A vector of column means}
\item{X1}{A matrix}
}
\description{
Internal function: Predict on new locatoins by MRTS method
}
\author{
Wen-Ting Wang
}
\keyword{internal}
