% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictMrtsRcppWithBasis}
\alias{predictMrtsRcppWithBasis}
\title{Internal function: Predict on new locatoins by MRTS method}
\usage{
predictMrtsRcppWithBasis(s, xobs_diag, s_new, BBBH, UZ, nconst, k)
}
\arguments{
\item{s}{A location matrix}

\item{xobs_diag}{A matrix of observations}

\item{s_new}{A new location matrix}

\item{BBBH}{A matrix for internal computing use}

\item{UZ}{A matrix for internal computing use}

\item{nconst}{A A vector of column means}

\item{k}{A rank}
}
\value{
A list of objects
\item{X}{A matrix}
\item{UZ}{A matrix}
\item{BBBH}{A matrix}
\item{nconst}{A vector of column means}
\item{X1}{A matrix}
}
\description{
Internal function: Predict on new locatoins by MRTS method
}
\keyword{internal}
