% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.mrts}
\alias{print.mrts}
\title{Internal function: print an mrts object}
\usage{
print.mrts(x, ...)
}
\arguments{
\item{x}{An mrts object}

\item{...}{Not used directly}
}
\description{
Internal function: print an mrts object
}
\keyword{internal}
