% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{removeWhitespace}
\alias{removeWhitespace}
\title{Internal function: remove white spaces for a given string}
\usage{
removeWhitespace(x)
}
\arguments{
\item{x}{A character}
}
\value{
A character
}
\description{
Internal function: remove white spaces for a given string
}
\keyword{internal}
