% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{selectBasis}
\alias{selectBasis}
\title{Internal function: select basis functions}
\usage{
selectBasis(
  data,
  loc,
  D = NULL,
  maxit = 50,
  avgtol = 1e-06,
  max_rank = NULL,
  sequence_rank = NULL,
  method = c("fast", "EM"),
  num_neighbors = 3,
  max_knot = 5000,
  DfromLK = NULL,
  Fk = NULL
)
}
\arguments{
\item{data}{An \emph{n} by \emph{T} data matrix (NA allowed) with
\eqn{z[t]} as the \emph{t}-th column.}

\item{loc}{\emph{n} by \emph{d} matrix of coordinates corresponding to \emph{n} locations.}

\item{D}{A diagonal matrix.}

\item{maxit}{An integer for the maximum number of iterations used in indeMLE.}

\item{avgtol}{A numeric for average tolerance used in indeMLE.}

\item{max_rank}{An integer of the maximum of K values.}

\item{sequence_rank}{An array of K values}

\item{method}{A character of a list of characters.}

\item{num_neighbors}{An integer.}

\item{max_knot}{An integer for the maximum number of knots}

\item{DfromLK}{An \emph{n} by \emph{n} diagonal matrix.}

\item{Fk}{An \emph{n} by \emph{K} matrix of basis function values with
each column being a basis function taken values at \code{loc}.}
}
\value{
An mrts object with 6 attributes
}
\description{
Internal function: select basis functions
}
\keyword{internal.}
