% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setDefaultAwght}
\alias{setDefaultAwght}
\title{Internal function: A wrapper of LatticeKrig::LKrigSetupAwght}
\usage{
setDefaultAwght(LKinfo)
}
\arguments{
\item{LKinfo}{LKinfo object}
}
\value{
list
}
\description{
Internal function: A wrapper of LatticeKrig::LKrigSetupAwght
}
\keyword{internal}
