% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkObject.R
\name{setLKnFRKOption}
\alias{setLKnFRKOption}
\title{Internal function: setLKnFRKOption}
\usage{
setLKnFRKOption(LK_obj, Fk, nc = NULL, Ks = NCOL(Fk), a.wght = NULL)
}
\arguments{
\item{LK_obj}{A list produced from \code{initializeLKnFRK}.}

\item{Fk}{An \emph{n} by \emph{K} matrix of basis function values with
each column being a basis function taken values at \code{location}.}

\item{nc}{A numeric made by \code{setNC}.}

\item{Ks}{An integer.}

\item{a.wght}{A numeric.}
}
\value{
list
}
\description{
Internal function: setLKnFRKOption
}
\keyword{internal}
