\name{setNC}
\alias{setNC}
\title{Internal function: set 'nc' for LKrigInfo}
\usage{
setNC(z, location, nlevel)
}
\arguments{
\item{z}{A matrix.}

\item{location}{A location matrix.}

\item{nlevel}{Number of levels in multi-resolution.}
}
\value{
A numeric value for \code{NC}.
}
\description{
Internal function: set \code{nc} for LKrigInfo.
}
\keyword{internal}
