% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkObject.R
\name{setRectangleAwght}
\alias{setRectangleAwght}
\title{Internal function: A wrapper of LatticeKrig::LKrigSetupAwght.LKRectangle}
\usage{
setRectangleAwght(LKinfo, ...)
}
\arguments{
\item{LKinfo}{LKinfo object}

\item{...}{Not used directly.}
}
\value{
list
}
\description{
Internal function: A wrapper of LatticeKrig::LKrigSetupAwght.LKRectangle
}
\keyword{internal}
