% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{shiftArray}
\alias{shiftArray}
\title{Internal function: A modifier of LatticeKrig::LKArrayShift}
\usage{
shiftArray(array_object, shift_index)
}
\arguments{
\item{array_object}{array}

\item{shift_index}{one-dim array}
}
\value{
A larray
}
\description{
Internal function: A modifier of LatticeKrig::LKArrayShift
}
\keyword{internal}
