% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{subKnot}
\alias{subKnot}
\title{Internal function: sampling knots}
\usage{
subKnot(x, nknot, xrng = NULL, nsamp = 1)
}
\arguments{
\item{x}{A matrix or an array}

\item{nknot}{A location matrix}

\item{xrng}{An array including two integers}
}
\value{
sampling knots
}
\description{
Internal function: sampling knots
}
\keyword{internal}
