% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{toBytes}
\alias{toBytes}
\title{Internal function: Convert value to bytes}
\usage{
toBytes(input_array)
}
\arguments{
\item{input_array}{An array consisting of the value and the unit}
}
\value{
A scalar in byte
}
\description{
Internal function: Convert value to bytes
}
\keyword{internal}
