% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{toSparseMatrix}
\alias{toSparseMatrix}
\title{Internal function: convert to a sparse matrix}
\usage{
toSparseMatrix(mat)
}
\arguments{
\item{mat}{A matrix or a dataframe}
}
\value{
sparse matrix
}
\description{
Internal function: convert to a sparse matrix
}
\keyword{internal}
