% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{wendland}
\alias{wendland}
\title{Internal function: A wendland function with k = 2 and l = 4}
\usage{
wendland(r)
}
\arguments{
\item{r}{A matrix of 2 or 3 d locations}
}
\value{
A matrix
}
\description{
Internal function: A wendland function with k = 2 and l = 4
}
\keyword{internal}
