% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cMLE}
\alias{cMLE}
\title{Internal function: estimate negative log-likelihood}
\usage{
cMLE(
  Fk,
  num_columns,
  sample_covariance_trace,
  inverse_square_root_matrix,
  matrix_JSJ,
  s = 0,
  ldet = 0,
  wSave = FALSE,
  onlylogLike = !wSave,
  vfixed = NULL
)
}
\arguments{
\item{Fk}{A  \emph{n} by \emph{K} matrix of basis function values with
each column being a basis function taken values at \code{loc}.}

\item{num_columns}{A positive numeric.}

\item{sample_covariance_trace}{A positive numeric.}

\item{inverse_square_root_matrix}{A matrix.}

\item{matrix_JSJ}{A multiplication matrix}

\item{s}{An integer. Sample size.}

\item{ldet}{An integer. Sample size.}

\item{wSave}{An integer. Sample size.}

\item{onlylogLike}{An integer. Sample size.}

\item{vfixed}{An integer. Sample size.}
}
\value{
A numeric.
}
\description{
Internal function: estimate negative log-likelihood
}
\keyword{internal.}
