% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{computeLikelihood}
\alias{computeLikelihood}
\title{Internal function: compute a negative log-likelihood}
\usage{
computeLikelihood(Data, Fk, M, s, Depsilon)
}
\arguments{
\item{Fk}{A  \emph{n} by \emph{K} matrix of basis function values with
each column being a basis function taken values at \code{loc}}

\item{M}{A \emph{K} by \emph{K} symmetric matrix}

\item{s}{A scalar}

\item{data}{A \emph{n} by \emph{T} data matrix (NA allowed) with
\eqn{z[t]} as the \emph{t}-th column.}

\item{Depsilion}{A \emph{n} by \emph{n} diagonal matrix}
}
\value{
A numeric
}
\description{
Internal function: compute a negative log-likelihood
}
\keyword{internal}
