% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{LKpeon}
\alias{LKpeon}
\title{Internal function: LKpeon}
\usage{
LKpeon(
  M,
  s,
  Fk,
  basis,
  weight,
  phi1,
  phi0,
  Q,
  lambda,
  phi0P,
  L = NULL,
  Data = NULL,
  only.wlk = FALSE,
  only.se = FALSE
)
}
\value{
A list.
}
\description{
Internal function: LKpeon
}
\keyword{internal}
