% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{neg2llik}
\alias{neg2llik}
\title{Internal function: estimate negative log-likelihood}
\usage{
neg2llik(d, s, v, sample_covariance_trace, n)
}
\arguments{
\item{d}{An array of nonnegative values.}

\item{s}{A positive numeric.}

\item{v}{A positive numeric.}

\item{sample_covariance_trace}{A positive numeric.}

\item{n}{An integer. Sample size.}
}
\value{
A numeric.
}
\description{
Internal function: estimate negative log-likelihood
}
\keyword{internal.}
