% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printor.R
\name{print.FRK}
\alias{print.FRK}
\title{Internal function: print an FRK object}
\usage{
\method{print}{FRK}(x, ...)
}
\arguments{
\item{x}{An FRK object}

\item{...}{Not used directly}
}
\description{
Internal function: print an FRK object
}
\keyword{internal}
