% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{selectBasis}
\alias{selectBasis}
\title{Internal function: select basis functions}
\usage{
selectBasis(
  data,
  loc,
  D = diag.spam(NROW(data)),
  maxit = 50,
  avgtol = 1e-06,
  maxK = NULL,
  Kseq = NULL,
  method = c("fast", "EM"),
  n.neighbor = 3,
  maxknot = 5000,
  DfromLK = NULL,
  Fk = NULL
)
}
\arguments{
\item{data}{A \emph{n} by \emph{T} data matrix (NA allowed) with
\eqn{z[t]} as the \emph{t}-th column.}

\item{loc}{\emph{n} by \emph{d} matrix of coordinates corresponding to \emph{n} locations.}

\item{D}{A diagonal matrix.}

\item{maxit}{An iteger for the maximum number of iterations used in indeMLE.}

\item{avgtol}{A numeric for average tolerance used in indeMLE.}

\item{maxK}{An integer of the maximum of K values.}

\item{Kseq}{An array of K values}

\item{method}{A character of a list of characters.}

\item{n.neighbor}{An integer.}

\item{maxknot}{An integer for the maximum number of knots}

\item{DfromLK}{A \emph{n} by \emph{n} diagonal matrix.}

\item{Fk}{A  \emph{n} by \emph{K} matrix of basis function values with
each column being a basis function taken values at \code{loc}.}
}
\value{
An mrts object with 6 attributes
}
\description{
Internal function: select basis functions
}
\keyword{internal.}
