% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkObject.R
\name{setNC}
\alias{setNC}
\title{Internal function: set 'nc' for LKrigInfo}
\usage{
setNC(z, location, nlevel)
}
\arguments{
\item{z}{A matrix}

\item{location}{A location matrix}

\item{nlelve}{An integer}
}
\value{
numeric
}
\description{
Internal function: set 'nc' for LKrigInfo
}
\keyword{internal}
